requirements.txt
coloredlogs==15.0.1
ffprobe_python==1.0.3
matplotlib==3.9.3
numpy==2.1.3
opencv_python==4.8.1.78
pandas==2.2.3
plotly==5.17.0
pycountry_convert==0.7.2
Requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
setuptools==68.0.0
Shapely==2.0.6
tqdm==4.66.1
xmltodict==0.13.0
